TLSConfig put out of dashboard
